add babel dependencies
add babel for make compatible app with all version of node
